Make network_tip and next_epoch optional
We should hear what Daedalus has to say about this before merging, but
I'm not sure what we could otherwise do.

If the node is not enough in-sync, we cannot know them.

Also replace `forecastFutureEpochStartUsingTip` with endTimeOfEpoch.

The need to forecast would arise from knowing the time delegation
statuses change (can be in epoch e+2), but this is only needed in
Shelley, and in Byron;Shelley, Shelley already has unlimited forecast range.

We do need endTimeOfEpoch to know the time of the next epoch when in
the Byron phase of Byron;Shelley. Getting the start time of the next
epoch won't work.